move query to components
