+++ frontend/src/components/Widgets/PaywallOverlay.jsx
@@ -3,6 +3,8 @@ import { useSelector } from "react-redux";
 import { proLink } from "../../config"
 
 function PaywallOverlay({ children, buttonText = 'Unlock with Lift Oracle Pro' }) {
+  const auth = useSelector((state) => state.auth);
+  console.log(auth);
   const isSubscribed = useSelector((state) => state.auth.isSubscribed)
   const enhancedChildren = isSubscribed 
     ? children
